feat: add results calculation with gpu (simple parallel FW) and convert to short

+++ Cuda/kernels.cuh
@@ -4,14 +4,14 @@
 
 
 
-__global__ void FW_simple_kernel(int* d_D, int n, int k) {
+__global__ void FW_simple_kernel(short* d_D, int n, int k) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     int j = blockIdx.y * blockDim.y + threadIdx.y;
   
     if (i < n && j < n) {
-        int ij = d_D[i * n + j];
-        int ik = d_D[i * n + k];
-        int kj = d_D[k * n + j];
+        short ij = d_D[i * n + j];
+        short ik = d_D[i * n + k];
+        short kj = d_D[k * n + j];
 
         if (ik + kj < ij)
             d_D[i * n + j] = ik + kj;
@@ -20,9 +20,9 @@ __global__ void FW_simple_kernel(int* d_D, int n, int k) {
 
 
 __forceinline__
-__device__ void blockedUpdateFW(int* C, int* A, int* B, int i, int j, const int blockSize){
+__device__ void blockedUpdateFW(short* C, short* A, short* B, int i, int j, const int blockSize){
     for(int k = 0; k < blockSize; k++){
-        int sum = A[i * blockSize + k] + B[k * blockSize + j];
+        short sum = A[i * blockSize + k] + B[k * blockSize + j];
         if(C[i * blockSize + j] > sum)
             C[i * blockSize + j] = sum;
         __syncthreads();
@@ -30,12 +30,12 @@ __device__ void blockedUpdateFW(int* C, int* A, int* B, int i, int j, const int 
 }
 
 // Aggiorna il blocco principale (k)
-__global__ void blocked_FW_phase1(int* d_D, int n, int k, const int blockSize){
+__global__ void blocked_FW_phase1(short* d_D, int n, int k, const int blockSize){
     int i = threadIdx.y;
     int j = threadIdx.x;
 
-    extern __shared__ int lmem[];
-    int* lmem_A = (int*)lmem;
+    extern __shared__ short lmem[];
+    short* lmem_A = (short*)lmem;
 
     lmem_A[i * blockSize + j] = d_D[(k * blockSize * n) + (k * blockSize) + (i * n + j)];
     __syncthreads();
@@ -47,7 +47,7 @@ __global__ void blocked_FW_phase1(int* d_D, int n, int k, const int blockSize){
 }
 
 // Aggiorna i blocchi nella stessa riga e colonna del blocco principale (k)
-__global__ void blocked_FW_phase2(int* d_D, int n, int k, const int blockSize){
+__global__ void blocked_FW_phase2(short* d_D, int n, int k, const int blockSize){
     // Seleziona l'indice (diagonale) da cui poi andremo a osservare
     // i blocchi nella stessa riga e colonna del blocco principale
     int x = blockIdx.x;
@@ -59,9 +59,9 @@ __global__ void blocked_FW_phase2(int* d_D, int n, int k, const int blockSize){
         return;
 
     // Uno per il blocco da modificare e l'altro per il blocco di dipendenza
-    extern __shared__ int lmem[];
-    int* lmem_A = (int*)lmem;
-    int* lmem_B = (int*)(&lmem_A[blockSize * blockSize]);
+    extern __shared__ short lmem[];
+    short* lmem_A = (short*)lmem;
+    short* lmem_B = (short*)(&lmem_A[blockSize * blockSize]);
 
     lmem_A[i * blockSize + j] = d_D[(x * blockSize * n) + (k * blockSize) + (i * n + j)];
     lmem_B[i * blockSize + j] = d_D[(k * blockSize * n) + (k * blockSize) + (i * n + j)];
@@ -85,7 +85,7 @@ __global__ void blocked_FW_phase2(int* d_D, int n, int k, const int blockSize){
 }
 
 // Aggiorna i blocchi restanti
-__global__ void blocked_FW_phase3(int* d_D, int n, int k, const int blockSize){
+__global__ void blocked_FW_phase3(short* d_D, int n, int k, const int blockSize){
     int x = blockIdx.y;
     int y = blockIdx.x;
 
@@ -95,10 +95,10 @@ __global__ void blocked_FW_phase3(int* d_D, int n, int k, const int blockSize){
     if(x == k && y == k)
         return;
 
-    extern __shared__ int lmem[];
-    int* lmem_A = (int*)lmem;
-    int* lmem_B = (int*)(&lmem_A[blockSize * blockSize]);
-    int* lmem_C = (int*)(&lmem_B[blockSize * blockSize]);
+    extern __shared__ short lmem[];
+    short* lmem_A = (short*)lmem;
+    short* lmem_B = (short*)(&lmem_A[blockSize * blockSize]);
+    short* lmem_C = (short*)(&lmem_B[blockSize * blockSize]);
 
     lmem_A[i * blockSize + j] = d_D[(x * blockSize * n) + (y * blockSize) + (i * n + j)];
     lmem_B[i * blockSize + j] = d_D[(x * blockSize * n) + (k * blockSize) + (i * n + j)];
